+++ sky/packages/sky/example/address_book/lib/main.dart
@@ -17,7 +17,7 @@ import 'package:sky/widgets/icon_button.dart';
 import 'package:sky/widgets/material.dart';
 import 'package:sky/widgets/navigator.dart';
 import 'package:sky/widgets/scaffold.dart';
-import 'package:sky/widgets/scrollable_viewport.dart';
+import 'package:sky/widgets/scrollable.dart';
 import 'package:sky/widgets/task_description.dart';
 import 'package:sky/widgets/theme.dart';
 import 'package:sky/widgets/tool_bar.dart';
