+++ sky/packages/sky/example/fitness/lib/main.dart
@@ -25,8 +25,7 @@ import 'package:sky/widgets/ink_well.dart';
 import 'package:sky/widgets/material.dart';
 import 'package:sky/widgets/radio.dart';
 import 'package:sky/widgets/scaffold.dart';
-import 'package:sky/widgets/scrollable_list.dart';
-import 'package:sky/widgets/scrollable_viewport.dart';
+import 'package:sky/widgets/scrollable.dart';
 import 'package:sky/widgets/snack_bar.dart';
 import 'package:sky/widgets/tool_bar.dart';
 import 'package:sky/widgets/card.dart';
