+++ sky/packages/sky/example/stocks/lib/main.dart
@@ -7,40 +7,39 @@ library stocks;
 import 'dart:math' as math;
 import 'dart:sky' as sky;
 
+import 'package:sky/editing/input.dart';
+import 'package:sky/painting/text_style.dart';
+import 'package:sky/rendering/box.dart';
 import 'package:sky/theme/colors.dart' as colors;
 import 'package:sky/theme/typography.dart' as typography;
 import 'package:sky/widgets/basic.dart';
+import 'package:sky/widgets/checkbox.dart';
 import 'package:sky/widgets/default_text_style.dart';
-import 'package:sky/widgets/navigator.dart';
-import 'package:sky/widgets/theme.dart';
-import 'package:sky/widgets/widget.dart';
-import 'package:sky/widgets/task_description.dart';
-import 'package:sky/rendering/box.dart';
-import 'package:sky/editing/input.dart';
-import 'package:sky/widgets/drawer.dart';
+import 'package:sky/widgets/dialog.dart';
 import 'package:sky/widgets/drawer_divider.dart';
 import 'package:sky/widgets/drawer_header.dart';
 import 'package:sky/widgets/drawer_item.dart';
+import 'package:sky/widgets/drawer.dart';
+import 'package:sky/widgets/flat_button.dart';
 import 'package:sky/widgets/floating_action_button.dart';
-import 'package:sky/widgets/icon.dart';
 import 'package:sky/widgets/icon_button.dart';
+import 'package:sky/widgets/icon.dart';
+import 'package:sky/widgets/ink_well.dart';
+import 'package:sky/widgets/material.dart';
 import 'package:sky/widgets/modal_overlay.dart';
+import 'package:sky/widgets/navigator.dart';
+import 'package:sky/widgets/popup_menu_item.dart';
 import 'package:sky/widgets/popup_menu.dart';
 import 'package:sky/widgets/radio.dart';
-import 'package:sky/widgets/snack_bar.dart';
 import 'package:sky/widgets/scaffold.dart';
-import 'package:sky/widgets/tabs.dart';
-import 'package:sky/widgets/tool_bar.dart';
-import 'package:sky/widgets/scrollable_list.dart';
-import 'package:sky/widgets/material.dart';
-import 'package:sky/widgets/checkbox.dart';
-import 'package:sky/widgets/popup_menu_item.dart';
-import 'package:sky/painting/text_style.dart';
-import 'package:sky/widgets/ink_well.dart';
-import 'package:sky/widgets/dialog.dart';
-import 'package:sky/widgets/flat_button.dart';
-import 'package:sky/widgets/scrollable_viewport.dart';
+import 'package:sky/widgets/scrollable.dart';
+import 'package:sky/widgets/snack_bar.dart';
 import 'package:sky/widgets/switch.dart';
+import 'package:sky/widgets/tabs.dart';
+import 'package:sky/widgets/task_description.dart';
+import 'package:sky/widgets/theme.dart';
+import 'package:sky/widgets/tool_bar.dart';
+import 'package:sky/widgets/widget.dart';
 
 //TODO(jimsimon): Is this needed?
 export 'package:sky/widgets/popup_menu.dart' show PopupMenuStatus;
